Ignore usepackage options when sorting in LaTeX
When sorting \usepackage declarations in LaTeX, one should only sort on
the package name, and not on the options. However, the options precede
the package name, so :sort includes the options by default.  This commit
ignores the options for any \usepackage when sorting.

For example:

\usepackage{amsmath}
\usepackage[utf8]{inputenc}

is now sorted in the order above, instead of as:

\usepackage[utf8]{inputenc}
\usepackage{amsmath}